+++ CHANGELOG.md
@@ -36,6 +36,10 @@
   and a state when channels don't exist. [#3531](https://github.com/GetStream/stream-chat-android/pull/3531)
 
 ### ✅ Added
+- Added `EventHandlingResult.WatchAndAdd` to results returned from `ChatEventHandler`.
+  After receiving the result, the SDK will call watch and adds the channel to the channels list. [#3569](https://github.com/GetStream/stream-chat-android/pull/3569)
+- Added handling `ChannelVisibleEvent`. Default `ChatEventHandler` will return `EventHandlingResult.WatchAndAdd`.
+  It can be customizing by overriding `ChatEventHandler::handleChannelVisibleEvent`. [#3569](https://github.com/GetStream/stream-chat-android/pull/3569)
 
 ### ⚠️ Changed
 
@@ -57,6 +61,7 @@
 - Fixed the way pagination scrolling worked for various non-core components (e.g. search, gallery/media/pinned message lists) [#3507](https://github.com/GetStream/stream-chat-android/pull/3507)
 - Added loading more indicator to PinnedMessageListView [#3507](https://github.com/GetStream/stream-chat-android/pull/3507)
 - Fix video scaling issue on the media preview screen. [#3560](https://github.com/GetStream/stream-chat-android/pull/3560)
+- Fixed refreshing `ChannelListView` after unhiding the channel. [#3569](https://github.com/GetStream/stream-chat-android/pull/3569)
 
 ### ⬆️ Improved
 
@@ -72,6 +77,7 @@
 ## stream-chat-android-compose
 ### 🐞 Fixed
 - Fix video scaling issue on the media preview screen. [#3560](https://github.com/GetStream/stream-chat-android/pull/3560)
+- Fixed refreshing `ChannelListView` after unhiding the channel. [#3569](https://github.com/GetStream/stream-chat-android/pull/3569)
 
 ### ⬆️ Improved